% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_metabolites.R
\name{compare_metabolites}
\alias{compare_metabolites}
\title{Comparison of metabolite sets between dynamics clusters of different experimental conditions}
\usage{
compare_metabolites(clusters, metabolite = "metabolite")
}
\arguments{
\item{clusters}{a dataframe or containing a column specifying the metabolite
names to be compared and cluster IDs (column named "cluster") of
clusters of similar dynamics, as well as a column "condition" specifying
the experimental conditions
to be compared}

\item{metabolite}{column in "clusters" that specifies either metabolite name
or KEGG ID or some other identifier}
}
\value{
a list holding a 1) dataframe of Jaccard indices between clusters
}
\description{
Uses the Jaccard Index to compare metabolite names between dynamics clusters of
different experimental conditions
}
\examples{
data("cluster")
comparison <- compare_metabolites(
  clusters = cluster
)
head(comparison[["Jaccard"]])
}
\seealso{
Visualization of metabolite similarity \code{\link[=heatmap_metabolites]{heatmap_metabolites()}}/
compare dynamics of clusters \code{\link[=compare_dynamics]{compare_dynamics()}}
}
