\name{data_sim}
\alias{data_sim}
\docType{data}
\title{
Dummy simulated data for practicing data documentation
}
\description{
Dummy simulated data to practice data documentation
}
\usage{data("data_sim")}
\format{
  Vector of length 1000 [0:1000]
}
\details{
created with data_sim <- sample(1000)
}
\source{
 NaN
}
\references{
NaN
}
\examples{
data_sim[1:10]
## maybe str(data_sim) ; plot(data_sim) ...
}
\keyword{datasets}
