% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{.calculate_distances}
\alias{.calculate_distances}
\title{compare_dynamics()}
\usage{
.calculate_distances(group_a, group_b, dynamics)
}
\arguments{
\item{group_a}{dataframe of one cluster of one condition}

\item{group_b}{dataframe of one cluster of a differenct condition than group_a}

\item{dynamics}{character vector specifying the columns that hold dynamic
estimates in data}
}
\value{
matrix of pairwise euclidean distances between two groups of vectors
}
\description{
compare_dynamics()
}
\keyword{internal}
