% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calculate_jaccard}
\alias{.calculate_jaccard}
\title{Function to calculate Jaccard index on two character vectors of metabolite
names}
\usage{
.calculate_jaccard(group_a, group_b)
}
\arguments{
\item{group_a}{group of clusters of metabolites}

\item{group_b}{group of clusters of metabolites}
}
\value{
the Jaccard index
}
\description{
Function to calculate Jaccard index on two character vectors of metabolite
names
}
\keyword{internal}
