% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu.R
\name{.eu}
\alias{.eu}
\title{euclidean distance
compare_dynamics()}
\usage{
.eu(a, b)
}
\arguments{
\item{a}{a numeric vector}

\item{b}{a numeric vector of same length as a}
}
\value{
euclidean distance between vectors
}
\description{
euclidean distance
compare_dynamics()
}
\keyword{internal}
