% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_boot_ph}
\alias{.get_boot_ph}
\title{get bootstraps for clustering of dynamics vectors (cluster_dynamics function)}
\usage{
.get_boot_ph(x, distance, agglomeration, B)
}
\arguments{
\item{x}{posterior of dynamics model}

\item{distance}{distance measure used for hierarchical clustering}

\item{agglomeration}{agglomeration method used for hierarchical clustering}

\item{B}{number of bootstraps}
}
\value{
bootstraps of clustering solution
}
\description{
get bootstraps for clustering of dynamics vectors (cluster_dynamics function)
}
\keyword{internal}
