% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module.R
\name{.get_module}
\alias{.get_module}
\title{internal helper function to retrieve experimental KEGG IDs annotated to module
get_ORA_annotations()}
\usage{
.get_module(M)
}
\arguments{
\item{M}{module name}
}
\value{
KEGG IDs of experimental metabolites annotated to module
}
\description{
internal helper function to retrieve experimental KEGG IDs annotated to module
get_ORA_annotations()
}
\keyword{internal}
