% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_background.R
\name{.get_module_background}
\alias{.get_module_background}
\title{internal helper function to retrieve background KEGG IDs of a module
get_ORA_annotations()}
\usage{
.get_module_background(M)
}
\arguments{
\item{M}{module name}
}
\value{
KEGG IDs of background metabolites annotated to module
}
\description{
internal helper function to retrieve background KEGG IDs of a module
get_ORA_annotations()
}
\keyword{internal}
