% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hierarchical_clustering}
\alias{.hierarchical_clustering}
\title{hierarchical clustering}
\usage{
.hierarchical_clustering(
  data,
  distance,
  agglomeration,
  minClusterSize,
  deepSplit
)
}
\arguments{
\item{data}{element "mu" of estimates}

\item{distance}{distance method}

\item{agglomeration}{agglomeration method of hierarchical clustering}

\item{minClusterSize}{minimum number of metabolites per of cluster \link[dynamicTreeCut]{cutreeDynamic}}

\item{deepSplit}{rough control over sensitivity of cluster analysis. Possible values are 0:4,
the higher the value, the more and smaller clusters will be produced by \link[dynamicTreeCut]{cutreeDynamic}}
}
\value{
list of input data including clustering solution, dendrogram, phylogram
}
\description{
hierarchical clustering
}
\keyword{internal}
