% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{.similarity}
\alias{.similarity}
\title{Jaccard Index: intersection/union
compare_metabolites()}
\usage{
.similarity(a, b)
}
\arguments{
\item{a}{a vector}

\item{b}{a vector}
}
\value{
Jaccard Index of a and b
}
\description{
Jaccard Index: intersection/union
compare_metabolites()
}
\keyword{internal}
