% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ORA_dataframes.R
\name{get_ORA_dataframes}
\alias{get_ORA_dataframes}
\title{Retrieve background and annotation information for over-representation analysis
(ORA)}
\usage{
get_ORA_dataframes(
  data,
  kegg = "KEGG",
  metabolite_name = "metabolite",
  update_background = FALSE
)
}
\arguments{
\item{data}{data set to be analyzed with ORA. Must at least contain a column
with KEGG IDs}

\item{kegg}{column name of "data" that holds KEGG IDs}

\item{metabolite_name}{column name of "data" that holds metabolite names}

\item{update_background}{logical. Should the background information be updated?
If TRUE this may take some time.}
}
\value{
a list with dataframes "background" and "annotation" needed for ORA
}
\description{
Uses the package KEGGREST to retrieve background and annotation information
needed for over-representation analysis. As KEGGREST only allows 10 queries
per second this might take some time to run, depending on the size of the
dataset and organism. The user should check afterwards if all functional modules are
applicable for the analysis of the dataset (p.e. organism, tissue).
}
\examples{
data("data_sim")
data_sim <- as.data.frame(SummarizedExperiment::colData(data_sim))
data <- data_sim[data_sim$metabolite == "ATP", ]
ORA_dataframes <- get_ORA_dataframes(data = data,
                                     kegg = "KEGG",
                                     metabolite_name = "metabolite",
                                     update_background = FALSE)
head(ORA_dataframes[["annotation"]])
}
\seealso{
\code{\link[=ORA_hypergeometric]{ORA_hypergeometric()}}
}
