% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_dynamics.R
\name{heatmap_dynamics}
\alias{heatmap_dynamics}
\title{plot bubble heatmap from the numerical fit of compare_dynamics()}
\usage{
heatmap_dynamics(
  estimates = metadata(data)[["comparison_dynamics"]][["estimates"]],
  data
)
}
\arguments{
\item{estimates}{dataframe of estimates of the mean distance
between clusters of different experimental conditions ("mean") and the
standard deviation ("sigma") obtain by function compare_dynamics()}

\item{data}{a dataframe or containing a column specifying the metabolite
names to be compared and cluster IDs (column named "cluster") of
clusters of similar dynamics, as well as a column "condition" specifying
the experimental conditions.
to be compared or a \link{SummarizedExperiment} storing the same information in
metadata(data) under "cluster"}
}
\value{
a bubble heat map where the color of the bubble represents the similarity
of two clusters in regards to their dynamics in the color and the size the
uncertainty of the similarity. Big bright bubbles mean high similarity with
low uncertainty.
}
\description{
plot bubble heatmap from the numerical fit of compare_dynamics()
}
\examples{
data("longitudinalMetabolomics")
longitudinalMetabolomics <- compare_dynamics(
  data = longitudinalMetabolomics,
  dynamics = c("mu1_mean", "mu2_mean", "mu3_mean", "mu4_mean"),
  cores = 1
)
heatmap_dynamics(data = longitudinalMetabolomics)
}
\seealso{
Do calculations for comparison of dynamics between clusters \code{\link[=compare_dynamics]{compare_dynamics()}}
}
