\name{IDs}
\alias{IDs}
\docType{data}
\title{
  KEGG ID mapping of metabolites in data set longitudinalMetabolomics
}
\usage{data("IDs")}
\format{
  A dataframe
  \describe{
    \item{\code{KEGG}}{KEGG ID of metabolites}
    \item{\code{replicate}}{column that specifies the measurement replicate}
  }
}