% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboDynamics-package.R
\docType{package}
\name{MetaboDynamics-package}
\alias{MetaboDynamics}
\alias{MetaboDynamics-package}
\title{MetaboDynamics: Bayesian analysis of longitudinal metabolomics data}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

MetaboDynamics is an R-package that provides a framework of probabilistic models to analyze longitudinal metabolomics data. It enables robust estimation of mean concentrations despite varying spread between timepoints and reports differences between timepoints as well as metabolite specific dynamics profiles that can be used for identifying "dynamics clusters" of metabolites of similar dynamics. Provides probabilistic over-representation analysis of KEGG functional modules and pathways as well as comparison between clusters of different experimental conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KatjaDanielzik/MetaboDynamics}
  \item Report bugs at \url{https://github.com/KatjaDanielzik/MetaboDynamics/issues}
}

}
\author{
\strong{Maintainer}: Katja Danielzik \email{katja.danielzik@uni-due.de} (\href{https://orcid.org/0009-0007-5021-6212}{ORCID})

Authors:
\itemize{
  \item Katja Danielzik \email{katja.danielzik@uni-due.de} (\href{https://orcid.org/0009-0007-5021-6212}{ORCID})
}

Other contributors:
\itemize{
  \item Simo Kitanovski (\href{https://orcid.org/0000-0003-2909-5376}{ORCID}) [contributor]
  \item Johann Matschke (\href{https://orcid.org/0000-0003-4878-8741}{ORCID}) [contributor]
  \item Daniel Hoffmann (\href{https://orcid.org/0000-0003-2973-7869}{ORCID}) [contributor]
}

}
\keyword{internal}
