% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{visualize clustering solution of cluster_dynamics()}
\usage{
plot_cluster(data)
}
\arguments{
\item{data}{result of \code{\link[=cluster_dynamics]{cluster_dynamics()}} function}
}
\description{
visualize clustering solution of cluster_dynamics()
}
\examples{
 data("longitudinalMetabolomics")
 plot_cluster(longitudinalMetabolomics)
}
