% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PPC.R
\name{plot_PPC}
\alias{plot_PPC}
\title{Plots posterior predictive check of numerical fit of Bayesian dynamics model}
\usage{
plot_PPC(posterior, data, scaled_measurement)
}
\arguments{
\item{posterior}{one dataframe per condition that
contains necessary information for Posterior predictive check
obtained by function diagnostics_dynamics()(named "PPC_condition")}

\item{data}{dataframe or colData of a SummarizedExperiment used to fit dynamics model}

\item{scaled_measurement}{column name of concentration values used to model fit, should be normalized by
experimental condition and metabolite to mean of zero and standard deviation
of one}
}
\value{
visual posterior predictive check
}
\description{
Plots posterior predictive check of numerical fit of Bayesian dynamics model
}
\examples{
data("longitudinalMetabolomics")
# only run after fit_dynamics_model(intra): see Vignette and documentation
# of function
longitudinalMetabolomics <- as.data.frame(SummarizedExperiment::colData(longitudinalMetabolomics))
data <- longitudinalMetabolomics[longitudinalMetabolomics$condition == "A" 
        & longitudinalMetabolomics$metabolite == "ATP", ]
fits <- fit_dynamics_model(
  data = data,
  scaled_measurement = "m_scaled", time = "time",
  condition = "condition", max_treedepth = 14,
  adapt_delta = 0.999, iter = 4000, cores = 1, chains = 1
)
diagnostics <- diagnostics_dynamics(
  data = data, iter = 4000, fits = fits,
  chains = 1, scaled_measurement = "m_scaled"
)
plot_PPC(posterior = diagnostics[["posterior_A"]], data = data, 
scaled_measurement = "m_scaled")
}
\seealso{
parent function \code{\link[=diagnostics_dynamics]{diagnostics_dynamics()}}
visualization function for diagnostics \code{\link[=plot_diagnostics]{plot_diagnostics()}}
}
