% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_metabolites.R
\name{compare_metabolites}
\alias{compare_metabolites}
\title{Comparison of metabolite sets between dynamics clusters of different experimental conditions}
\usage{
compare_metabolites(data, metabolite = "metabolite")
}
\arguments{
\item{data}{a dataframe or containing a column specifying the metabolite
names to be compared and cluster IDs (column named "cluster") of
clusters of similar dynamics, as well as a column "condition" specifying
the experimental conditions.
to be compared or a \link[SummarizedExperiment]{SummarizedExperiment} storing the same information in
metadata(data) under "cluster"}

\item{metabolite}{column in "data" that specifies either metabolite name
or KEGG ID or some other identifier}
}
\value{
a dataframe of Jaccard indices between data or if data input
was a SummarizedExperiment results are stored in metadata(data) under
"comparison_metabolites"
}
\description{
Uses the Jaccard Index to compare metabolite names between dynamics clusters of
different experimental conditions
}
\examples{
data("longitudinalMetabolomics")
longitudinalMetabolomics <- compare_metabolites(
  data = longitudinalMetabolomics
)
S4Vectors::metadata(longitudinalMetabolomics)[["comparison_metabolites"]]
}
\seealso{
Visualization of metabolite similarity \code{\link[=heatmap_metabolites]{heatmap_metabolites()}}/
compare dynamics of clusters \code{\link[=compare_dynamics]{compare_dynamics()}}
}
