% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_metabolites.R
\name{compare_metabolites}
\alias{compare_metabolites}
\title{Comparison of metabolites between clusters of different experimental conditions}
\usage{
compare_metabolites(clusters)
}
\arguments{
\item{clusters}{a dataframe containing the columns "metabolite" specifying the
metabolite names to be compared and cluster IDs(column named "cluster") of
clusters of similar dynamics,as well as a column "condition" specifying
the experimental conditions
to be compared}
}
\value{
a list holding a 1) dataframe of Jaccard indices between clusters
2) a ggplot2 object visualizing the cluster comparison
}
\description{
Uses the Jaccard Index to compare metabolite names between clusters of
different experimental conditions
}
\examples{
\dontrun{
compare_metabolites <- functions(
  clusters = cluster
)
}
}
\keyword{internal}
