% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_dynamics.R
\name{diagnostics_dynamics}
\alias{diagnostics_dynamics}
\title{Extracts diagnostic criteria from numeric fit of Bayesian model of dynamics}
\usage{
diagnostics_dynamics(
  data,
  N = nrow(data),
  M = length(unique(data$metabolite)),
  t = length(unique(data$time)),
  iter = 2000,
  warmup = iter/4,
  chains = 4,
  fits,
  scaled_measurement = "m_scaled"
)
}
\arguments{
\item{data}{dataframe or colData of a SummarizedExperiment used to fit dynamics model
column of "time" that contains time as numeric, make sure your
time column is ordered from lowest to highest for the model to work}

\item{N}{number of rows in that dataframe}

\item{M}{number of metabolites in experimental dataset}

\item{t}{number of timepoints in experimental dataset}

\item{iter}{number of iterations used for model fit}

\item{warmup}{number of warmup iterations used for model fit}

\item{chains}{number of chains used for model fit}

\item{fits}{list of model fits for which diagnostics should be extracted, is the
object that gets returned by fit_dynamics_model()}

\item{scaled_measurement}{concentration values used to model fit, should be normalized by
experimental condition and metabolite to mean of zero and standard deviation
of one}
}
\value{
a list which contains diagnostics criteria of all conditions in a
dataframe (named "model_diagnostics") and one dataframe per condition that
contains necessary information for Posterior predictive check
(named "PPC_condition").
}
\description{
gathers number of divergences, rhat values, number of effective
samples (n_eff) and provides plots for diagnostics criteria as well as
posterior predictive checks. Dataframe "model_diagnostics" contains
information about experimental condition, number of divergent transitions
and rhat and neff values for all timepoints.
}
\examples{
data("longitudinalMetabolomics")
# only run after fit_dynamics_model(intra): see Vignette and documentation
# of function
longitudinalMetabolomics <- as.data.frame(SummarizedExperiment::colData(longitudinalMetabolomics))
data <- longitudinalMetabolomics[longitudinalMetabolomics$condition == "A" &
  longitudinalMetabolomics$metabolite == "ATP", ]
fits <- fit_dynamics_model(
  data = data,
  scaled_measurement = "m_scaled", time = "time",
  condition = "condition", max_treedepth = 14,
  adapt_delta = 0.999, iter = 4000, cores = 1, chains = 1
)
diagnostics <- diagnostics_dynamics(
  data = data, iter = 4000, fits = fits,
  chains = 1, scaled_measurement = "m_scaled"
)
head(diagnostics[["model_diagnostics"]])
head(diagnostics[["posterior_A"]])
}
\seealso{
\code{\link[=estimates_dynamics]{estimates_dynamics()}}
parent function \code{\link[=fit_dynamics_model]{fit_dynamics_model()}}
visualization functions: \code{\link[=plot_diagnostics]{plot_diagnostics()}}/\code{\link[=plot_PPC]{plot_PPC()}}
}
