% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_fit_dynamics_input}
\alias{.check_fit_dynamics_input}
\title{input checks for fit_dynamics_model}
\usage{
.check_fit_dynamics_input(
  model,
  data,
  metabolite,
  time,
  condition,
  scaled_measurement,
  counts,
  assay,
  chains,
  cores,
  adapt_delta,
  max_treedepth,
  iter,
  warmup
)
}
\description{
input checks for fit_dynamics_model
}
\keyword{internal}
