% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_boot_ph}
\alias{.get_boot_ph}
\title{get bootstrapps for clustering of dynamics vectors (cluster_dynamics function)}
\usage{
.get_boot_ph(x, distance, agglomeration, B)
}
\description{
get bootstrapps for clustering of dynamics vectors (cluster_dynamics function)
}
\keyword{internal}
