% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_dynamics.R
\name{estimates_dynamics}
\alias{estimates_dynamics}
\title{Extracts parameter estimates from numeric fit of Bayesian model of dynamics}
\usage{
estimates_dynamics(
  data,
  M = length(unique(data$metabolite)),
  t = length(unique(data$time)),
  kegg = "KEGG",
  condition = "dose",
  fits,
  iter = 2000,
  warmup = iter/4,
  chains = 4,
  samples = 1
)
}
\arguments{
\item{data}{dataframe or colData of a SummarizedExperiment used used to fit dynamics model}

\item{M}{number of metabolites, default requires a column in data named "metabolite"}

\item{t}{number of unique timepoints in data, the default requires a column named "time"}

\item{kegg}{column in "data" that contains the KEGG IDs or other identifier of metabolites}

\item{condition}{name of column in dataframe data that specifies the experimental condition, default requires a column named "dose"}

\item{fits}{list of model fits for which estimates should be extracted}

\item{iter}{how many iterations were used to fit the dynamics model}

\item{warmup}{how many warm-up iterations were used to fit the dynamics model}

\item{chains}{how many chains were used to fit the dynamics model}

\item{samples}{how many posterior samples should be drawn (p.e. for check of clustering precision)}
}
\value{
a list of dataframes (one per experimental condition) that contains
the estimates at the timepoints and samples from the posterior
(number as specified in samples)
}
\description{
Extracts the mean concentrations (mu) at every timepoint from the dynamics model fit, the 95\% highest density interval (HDI), the estimated standard deviation of metabolite concentrations at every time point (sigma), and the pooled standard deviation of every metabolite over all timepoints (lambda).
Additionally samples from the posterior of mu can be drawn. This can be helpful if p.e. one wants to estimate the clustering precision. Lambda can be used for clustering algorithms such as VSClust that also take the variance into account.
}
\examples{
data("data_sim")
data_sim <- as.data.frame(SummarizedExperiment::colData(data_sim))
data <- data_sim[data_sim$condition == "A" & data_sim$metabolite == "ATP", ]
fits <- fit_dynamics_model(
  data = data,
  scaled_measurement = "m_scaled", condition = "condition",
  max_treedepth = 14, adapt_delta = 0.999, iter = 4000, cores = 1, chains = 1
)
estimates <- estimates_dynamics(
  data = data, fits = fits, iter = 4000,
  chains = 1, condition = "condition"
)
head(estimates)

}
\seealso{
\code{\link[=fit_dynamics_model]{fit_dynamics_model()}}/\code{\link[=diagnostics_dynamics]{diagnostics_dynamics()}}
}
