% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_diagnostics_dynamics.R
\name{extract_diagnostics_dynamics}
\alias{extract_diagnostics_dynamics}
\title{Extracts diagnostic criteria from dynamics model}
\usage{
extract_diagnostics_dynamics(
  data,
  N = nrow(data),
  M = length(unique(data$metabolite)),
  t = length(unique(data$time)),
  iter = 2000,
  warmup = iter/4,
  chains = 4,
  fits,
  cpc = "log_cpc_stand"
)
}
\arguments{
\item{data}{dataframe used to fit dynamics model}

\item{N}{number of rows in used dataframe}

\item{M}{number of metabolites in experimental dataset}

\item{t}{number of timepoints in experimental dataset}

\item{iter}{number of iterations used for model fit}

\item{warmup}{number of warmup iterations used for model fit}

\item{chains}{number of chains used for model fit}

\item{fits}{list of models for which diagnostics should be extracted, is the
object that gets returned by fit_dynamics_model()}

\item{cpc}{concentration values used to model fit, should be normalized by
experimental condition and metabolite to mean of zero and standard deviation
of one}
}
\value{
a list which contains diagnostics criteria of all conditions in a
dataframe (named "model_diagnostics") and one dataframe per condition that
contains necessary information for Posterior predictive check
(named "PPC_condition"). Additionally plots for diagnostics and PPC named
"plot_criteria" and "plot_PPC_condition".
}
\description{
gathers number of divergences, rhat values, number of effective
samples (n_eff) and provides plots for diagnostics criteria as well ad
Posterior predictive checks. Dataframe "model_diagnostics" contains
information about experimental condition, number of divergent transitions
and rhat and neff values for all timepoints
}
\examples{
\dontrun{
data("intra")
# only run after fit_dynamics_model(intra): see Vignette and documentation
# of function
extract_diagnostics_dynamics(
  data = intra, max_treedepth = 14,
  adapt_delta = 0.999, iter = 4000, fits = fits_dynamics
)
}
}
