% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dynamics_model.R
\name{fit_dynamics_model}
\alias{fit_dynamics_model}
\title{Fits dynamics model}
\usage{
fit_dynamics_model(
  data,
  metabolite = "metabolite",
  time = "time",
  condition = "dose",
  scaled_measurement = "m_scaled",
  chains = 4,
  cores = 4,
  adapt_delta = 0.95,
  max_treedepth = 10,
  iter = 2000,
  warmup = iter/4
)
}
\arguments{
\item{data}{concentration table containing the columns "metabolite",
"condition", "cpc_stand" by default or colData of a SummarizedExperiment
\linkS4class{SummarizedExperiment} object}

\item{metabolite}{column of "data" that contains the metabolite names or IDs}

\item{time}{column of "time" that contains time as numeric, make sure your
time column is ordered from lowest to highest for the model to work}

\item{condition}{column of "data" that contains the experimental conditions}

\item{scaled_measurement}{column of "data" that contains the concentrations per cell,
centered and normalized per metabolite and experimental condition (mean=0, sd=1)}

\item{chains}{how many Markov-Chains should be used for model fitting, use at
least two, default=4}

\item{cores}{how many cores should be used for model fitting; this
parallelizes the model fitting and therefore speeds it up; default=4}

\item{adapt_delta}{target average acceptance probability, can be adapted if
divergent transitions are reported, default is 0.95}

\item{max_treedepth}{can be adapted if model throws warnings about hitting
max_treedepth, warnings are mostly efficiency not validity concerns and
treedepth can be raised, default=10}

\item{iter}{how many iterations are run, increasing might help with effective}

\item{warmup}{how many iterations the model warms up for, increasing might
facilitate efficiency, must be at least 25\% of ITER, default=iter/4}
}
\value{
returns a list of model fits. One model fit named fit_condition per
experimental condition
}
\description{
Employs a hierarchical model that assumes a normal distribution of
standardized (mean=0, sd=1) log(cpc) (cpc = concentration per cell) values for robust estimation of mean
concentrations over time of single metabolites at single experimental
conditions.
}
\examples{
data("data_sim")
data_sim <- as.data.frame(SummarizedExperiment::colData(data_sim)
data <- data_sim[data_sim$condition == "A" & data_sim$metabolite == "ATP", ]
fits <- fit_dynamics_model(
  data = data,
  scaled_measurement = "m_scaled", condition = "condition", 
  max_treedepth = 14, adapt_delta = 0.999, iter = 4000, cores = 1, chains = 1
)
fits

}
\seealso{
\code{\link[=extract_diagnostics_dynamics]{extract_diagnostics_dynamics()}}/\code{\link[=extract_estimates_dynamics]{extract_estimates_dynamics()}}
}
