% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_dynamics.R
\name{heatmap_dynamics}
\alias{heatmap_dynamics}
\title{plot bubble heatmap from the numerical fit of compare_dynamics()}
\usage{
heatmap_dynamics(estimates, clusters)
}
\arguments{
\item{estimates}{dataframe of estimates of the mean distance
between #' clusters of different experimental conditions ("mean") and the
standard deviation ("sigma")}

\item{clusters}{a dataframe containing the dynamics and
cluster IDs(column named "cluster") of clusters of similar dynamics,
as well as a column "condition" specifying the experimental conditions
to be compared.}
}
\value{
a bubble heatmap where the color of the bubble represents the similarity
of two clusters in regards to their dynamics in the color and the size the
uncertainty of the similarity. Big bright bubbles mean high similarity with
low uncertainty.
}
\description{
plot bubble heatmap from the numerical fit of compare_dynamics()
}
\examples{
data("cluster")
# fit model 
comparison <- compare_dynamics(
  clusters = cluster,
  dynamics = c("mu1_mean", "mu2_mean", "mu3_mean", "mu4_mean"),
  cores = 1
)
plot <- heatmap_dynamics(estimates=comparison[["estimates"]],clusters=cluster)
plot
}
\seealso{
\code{\link[=compare_dynamics]{compare_dynamics()}}
}
