% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_metabolites.R
\name{heatmap_metabolites}
\alias{heatmap_metabolites}
\title{plot heatmap from comparison of metabolite composition compare_metabolites()}
\usage{
heatmap_metabolites(distances, clusters)
}
\arguments{
\item{distances}{dataframe of Jaccard indices between clusters obtained
by function compare_metabolites()}

\item{clusters}{a dataframe containing the columns "metabolite" specifying the
metabolite names to be compared and cluster IDs(column named "cluster") of
clusters of similar dynamics, as well as a column "condition" specifying
the experimental conditions
to be compared}
}
\value{
a heatmap where the color of the tile represents the similarity
of two clusters in regards to their metabolite composition. The brighter
the color the more similar the metabolite compositions.
}
\description{
plot heatmap from comparison of metabolite composition compare_metabolites()
}
\examples{
data("cluster")
comparison <- compare_metabolites(
  clusters = cluster
)
heatmap_metabolites(distances=comparison[["Jaccard"]],clusters=cluster)
}
\seealso{
\code{\link[=compare_metabolites]{compare_metabolites()}}
}
