\name{IDs}
\alias{IDs}
\docType{data}
\title{
  KEGG ID mapping of metabolites in data set longitudinalMetabolomics
}
\description{
Example data set of KEGG ID annotations of metabolites as needed for \code{\link[=ORA_hypergeometric]{ORA_hypergeometric()}}
}
\usage{data("IDs")}
\format{
  A dataframe
  \describe{
    \item{\code{KEGG}}{KEGG ID of metabolites}
    \item{\code{replicate}}{column that specifies the measurement replicate}
  }
}
\keyword{datasets}
