% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboDynamics-package.R
\docType{package}
\name{MetaboDynamics-package}
\alias{MetaboDynamics}
\alias{MetaboDynamics-package}
\title{MetaboDynamics: Bayesian Analysis of longitudinal metabolomics data}
\description{
MetaboDynamics: framework of Bayesian models to analyze longitudinal metabolomics data
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\author{
\strong{Maintainer}: Katja Danielzik \email{katja.danielzik@uni-due.de} (\href{https://orcid.org/0009-0007-5021-6212}{ORCID})

}
\keyword{internal}
