% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboDynamics-package.R
\docType{package}
\name{MetaboDynamics-package}
\alias{MetaboDynamics}
\alias{MetaboDynamics-package}
\title{MetaboDynamics: Bayesian Analysis of longitudinal metabolomics data}
\description{
MetaboDynamics is an R-package that provides a framework of probabilistic models to analyze longitudinal metabolomics data. It enables robust estimation of mean concentrations despite varying spread between timepoints and reports differences between timepoints as well as metabolite specific dynamic profiles that can be used for clustering. Probabilistic Over-representation analysis and comparison betweeen clusters of different experimental conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KatjaDanielzik/MetaboDynamics}
  \item Report bugs at \url{https://github.com/KatjaDanielzik/MetaboDynamics/issues}
}

}
\author{
\strong{Maintainer}: Katja Danielzik \email{katja.danielzik@uni-due.de} (\href{https://orcid.org/0009-0007-5021-6212}{ORCID})

Other contributors:
\itemize{
  \item Simo Kitanovski [contributor]
  \item Johann Matschke [contributor]
  \item Daniel Hoffmann [contributor]
}

}
\keyword{internal}
