% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{visualize clustering solution of cluster_dynamics()}
\usage{
plot_cluster(data)
}
\arguments{
\item{data}{result of \code{\link[=cluster_dynamics]{cluster_dynamics()}} function: either a list of data frames or a SummarizedExperiment object}
}
\value{
a list of plots. Per experimental condition: 1) a 'bubbletree': a phylogram with numbers on nodes
indicating in how many bootstraps of the posterior estimates the same clustering
solution was generated, 2) cluster affiliation of metabolites, 3) dynamics of metabolites per cluster,
4) patchwork of 1-3, 5) order of the tips in bubbletree: needed for matching cluster plots and ORA
}
\description{
visualize clustering solution of cluster_dynamics()
}
\examples{
data("longitudinalMetabolomics")
plots <- plot_cluster(longitudinalMetabolomics[, longitudinalMetabolomics$condition == "A"])

plots[["trees"]][["A"]]
}
\seealso{
\code{\link[=cluster_dynamics]{cluster_dynamics()}}
}
