% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plot diagnostic criteria of numerical fit of Bayesian model of dynamics}
\usage{
plot_diagnostics(
  diagnostics,
  data,
  divergences = TRUE,
  max_treedepth = TRUE,
  Rhat = TRUE,
  n_eff = TRUE
)
}
\arguments{
\item{diagnostics}{dataframe containing diagnostics criteria from the numerical
fit of Bayesian model of dynamics obtained by function diagnostics_dynamics()}

\item{data}{dataframe or colData of a SummarizedExperiment used to fit dynamics model
must contain column "time"}

\item{divergences}{should number of divergent transitions be visualized?}

\item{max_treedepth}{should number of exeeded maximum treedepth be visualized?}

\item{Rhat}{should Rhat be visualized?}

\item{n_eff}{should number of effective samples be visualized?}
}
\value{
plots of diagnostic criteria of numerical fit of Bayesian model of
dynamics
}
\description{
Plot diagnostic criteria of numerical fit of Bayesian model of dynamics
}
\examples{
data("longitudinalMetabolomics")
# only run after fit_dynamics_model(intra): see Vignette and documentation
# of function
longitudinalMetabolomics <- as.data.frame(SummarizedExperiment::colData(longitudinalMetabolomics))
data <- longitudinalMetabolomics[longitudinalMetabolomics$condition == "A" &
  longitudinalMetabolomics$metabolite == "ATP", ]
fits <- fit_dynamics_model(
  data = data,
  scaled_measurement = "m_scaled", time = "time",
  condition = "condition", max_treedepth = 14,
  adapt_delta = 0.999, iter = 4000, cores = 1, chains = 1
)
diagnostics <- diagnostics_dynamics(
  data = data, iter = 4000, fits = fits,
  chains = 1, scaled_measurement = "m_scaled"
)
plot_diagnostics(diagnostics = diagnostics[["model_diagnostics"]], data = data)
}
\seealso{
parent function \code{\link[=diagnostics_dynamics]{diagnostics_dynamics()}}
visualization function for posterior predictive check /\code{\link[=plot_PPC]{plot_PPC()}}
}
