% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Visualization of parameter estimates from numeric fit of Bayesian model of dynamics}
\usage{
plot_estimates(estimates, data, delta_t = TRUE, dynamics = TRUE)
}
\arguments{
\item{estimates}{a list of dataframes (one per experimental condition) that contains
the estimates at the timepoints and samples from the posterior generated
by estimates_dynamics()}

\item{data}{dataframe or colData of a SummarizedExperiment used used to fit dynamics model}

\item{delta_t}{should differences between timepoints be plotted?}

\item{dynamics}{should dynamics be plotted?}
}
\value{
Visualization of differences between timepoints(delta_t) and dynamics
profiles of single metabolites
}
\description{
Visualization of parameter estimates from numeric fit of Bayesian model of dynamics
}
\examples{
data("longitudinalMetabolomics")
longitudinalMetabolomics <- as.data.frame(SummarizedExperiment::colData(longitudinalMetabolomics))
data <- longitudinalMetabolomics[longitudinalMetabolomics$condition == "A" 
   & longitudinalMetabolomics$metabolite == "ATP", ]
fits <- fit_dynamics_model(
  data = data,
  scaled_measurement = "m_scaled", condition = "condition",
  max_treedepth = 14, adapt_delta = 0.999, iter = 4000, cores = 1, chains = 1
)
estimates <- estimates_dynamics(
  data = data, fits = fits, iter = 4000,
  chains = 1, condition = "condition"
)
plot_estimates(estimates = estimates, data = data, delta_t=TRUE)
plot_estimates(estimates = estimates, data = data, dynamics=TRUE)
}
\seealso{
\code{\link[=estimates_dynamics]{estimates_dynamics()}}
}
