% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ORA.R
\name{plot_ORA}
\alias{plot_ORA}
\title{Plot results of over-representation analysis with ORA_hypergeometric()}
\usage{
plot_ORA(ORA)
}
\arguments{
\item{ORA}{result dataframe from over-representation analysis with ORA_hypergeometric}
}
\value{
a plot of the over-representation analysis
}
\description{
Plot results of over-representation analysis with ORA_hypergeometric()
}
\examples{
data("cluster")
head(cluster)
data("modules_compounds")
head(modules_compounds)
data("metabolite_modules")
head(metabolite_modules)
ORA <- ORA_hypergeometric(
  background = modules_compounds,
  annotations = metabolite_modules, clusters = cluster,
  tested_column = "middle_hierarchy"
)
plot_ORA(ORA)
}
\seealso{
function to do ORA \code{\link[=ORA_hypergeometric]{ORA_hypergeometric()}}
}
